Small fixes

+++ src/ORMMapViewer/Model/Data/TangramData.cs
@@ -28,7 +28,7 @@ namespace ORMMap.Model.Data
 		}
 
 		// mapZoomLevels from 1 to 16 inclusive
-		public override double ConvertToMapZoom(double zoom)
+		public override int ConvertToMapZoom(double zoom)
 		{
 			return Math.Min(16, Math.Floor(zoom));
 		}
